// src/modules/permissions/enums/controller-names.enum.ts
export enum ControllerNames {
    UsersController = 'UsersController',
    MeController = 'MeController',
    PostsController = 'PostsController',
    TagsController = 'TagsController',
    CommentsController = 'CommentsController',
    MetaOptionsController = 'MetaOptionsController',
    UploadsController = 'UploadsController',
    RolesController = 'RolesController',
    PermissionsController = 'PermissionsController',
}